aqua-assessments/requirements.txt
fastapi~=0.110.0
modal~=0.62
pandas~=1.5.0
pydantic~=1.10.0
pytest~=8.0.0
requests~=2.31.0
numpy~=1.26.4
sqlalchemy~=1.4.0
psycopg2~=2.9.0
python-dotenv
nltk~=3.6.0